+++ src/uniqdiff/__init__.py
@@ -54,7 +54,7 @@ from uniqdiff.exceptions import (
 from uniqdiff.fields import FieldChange, FieldDiffResult, FieldDiffStats, iter_field_diff_rows
 from uniqdiff.fuzzy import compare_fuzzy_strings
 from uniqdiff.normalizers import compose_normalizers, string_normalizer
-from uniqdiff.output import iter_result_rows, iter_result_values
+from uniqdiff.output import iter_event_rows, iter_result_rows, iter_result_values
 from uniqdiff.result import CompareResult, CompareStats
 from uniqdiff.schema import ColumnSchema, SchemaDiffResult, SchemaResult
 from uniqdiff.streaming import iter_sorted_diff, write_sorted_diff
@@ -112,6 +112,7 @@ __all__ = [
     "infer_schema",
     "intersection",
     "iter_compare_events",
+    "iter_event_rows",
     "iter_result_rows",
     "iter_result_values",
     "iter_field_diff_rows",